requirements.txt
APScheduler==3.6.3
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
decorator==4.4.2
html5lib==1.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
python-telegram-bot==13.0
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
webencodings==0.5.1
wrapt==1.12.1
